feat: all C and assembly files compile cleanly for x86_64 NOTE: kernel does not boot yet - 64-bit migration in progress - start.asm still needs rewriting for 64-bit - Two stage bootloader needed - 4-level paging needed - Higher half kernel needed

+++ kernel/mem.c
@@ -47,7 +47,7 @@ void mem_init()
 
 // --- kmalloc --------------------------------------
 
-void* kmalloc(uint32_t size)
+void* kmalloc(uint64_t size)
 {
     // Allocating zero bytes makes no sense
     if(size == 0)
@@ -119,7 +119,7 @@ void* kmalloc(uint32_t size)
     int count = 0;
     while(dbg && count < 5) {
         kserial_printf("  block %d: size=%d free=%d next=0x%x\n",
-            count, dbg->size, dbg->free, (uint32_t)dbg->next);
+            count, dbg->size, dbg->free, (uint64_t)dbg->next);
         dbg = dbg->next;
         count++;
     }
